// 04-IMPL/mejoredu-portal/src/app/follow/mir/interfaces/mir-data.interface.ts
export interface MIRDataI {
  id: number;
  mir: string;
  indicador: string;
  unidad_1: string;
  programado_1: number;
  porcentaje_programado_1: string;
  reportado_1: number;
  porcentaje_reportado_1: string;
  unidad_2: string;
  programado_2: number;
  porcentaje_programado_2: string;
  reportado_2: number;
  porcentaje_reportado_2: string;
  unidad_3: string;
  programado_3: number;
  porcentaje_programado_3: string;
  reportado_3: number;
  porcentaje_reportado_3: string;
  unidad_4: string;
  programado_4: number;
  porcentaje_programado_4: string;
  reportado_4: number;
  porcentaje_reportado_4: string;
  alcanzadoAcumulado: number;
  porcentajeAcumulado: string;
  estatus: string;
}